DNA sequence alignments of centromeres and flanking DNA extracted from SGRP
gff files give coordinates of CDEI, CDEII and CDEIII for each
alignment. The SGRP PALAS alignments analysed in Liti et al (2009) for
the 16 centromere loci of 35 strains of S. paradoxus were extracted
using alicat.pl (a perl script provided with the SGRP data). A quality
threshold of Q40 was used for the SGRP data, and sites with a lower
quality score were represented with an “N”.